% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helperFuns.R
\name{storeInDB}
\alias{storeInDB}
\title{storeInDB a function to help transmit the key/value pairs to the redis server}
\usage{
storeInDB(key, value)
}
\arguments{
\item{key}{the redis key}

\item{value}{the value to store}
}
\description{
storeInDB a function to help transmit the key/value pairs to the redis server
}

